+++ spinta/commands/__init__.py
@@ -932,3 +932,29 @@ def inspect(*args) -> Iterator[ManifestSchema]:
     comparing source names.
 
     """
+
+
+@overload
+def merge(
+    context: Context,
+    manifest: Manifest,
+    old: Any,
+    new: Any,
+) -> None:
+    """Merge new manifest into old."""
+
+
+@overload
+def merge(
+    context: Context,
+    manifest: Manifest,
+    old: Any,
+    new: Any,
+    has_manifest_priority: bool
+) -> None:
+    """Merge new manifest into old."""
+
+
+@command()
+def merge(*args) -> None:
+    """Merge new manifest into old."""