+++ spinta/exceptions.py
@@ -562,3 +562,15 @@ class RemoteClientScopesNotGiven(RemoteClientError):
 
 class RequiredProperty(UserError):
     template = "Property is required."
+
+
+class UnableToCast(UserError):
+    template = "Unable to cast {value} to {type} type."
+
+
+class ReferencedObjectNotFound(UserError):
+    template = "Referenced object {id!r} not found."
+
+
+class ReferringObjectFound(UserError):
+    template = "Object {id!r} is still being referenced in table {model!r}"