+++ spinta/types/model.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import List
 from typing import Optional
-from typing import Tuple
+from typing import Union
 from typing import overload
 
 import itertools
@@ -44,6 +44,7 @@ from spinta.types.namespace import load_namespace_from_name
 from spinta.units.helpers import is_unit
 from spinta.utils.enums import enum_by_value
 from spinta.utils.schema import NA
+from spinta.types.datatype import Ref
 
 if TYPE_CHECKING:
     from spinta.datasets.components import Attribute
@@ -231,7 +232,11 @@ def load(
     else:
         prop.external = NA
     commands.load(context, prop.dtype, data, manifest)
-
+    if prop.model.unique:
+        if isinstance(prop.dtype, Ref):
+            if '.id' not in prop.name:
+                prop.model.unique = [list(map(lambda val: val.replace(
+                    prop.name, prop.name + '._id'), val)) for val in prop.model.unique]
     unit: Optional[str] = prop.enum
     if unit is None:
         prop.given.enum = None
@@ -249,14 +254,17 @@ def load(
     return prop
 
 
-def load_level(component: Component, given_level):
+def load_level(
+    component: Component,
+    given_level: Union[Level, int, str],
+):
     if given_level:
         if isinstance(given_level, Level):
             level = given_level
         else:
-            if given_level.isnumeric():
+            if isinstance(given_level, str) and given_level.isdigit():
                 given_level = int(given_level)
-            else:
+            if not isinstance(given_level, int):
                 raise InvalidLevel(component, level=given_level)
             level = enum_by_value(component, 'level', Level, given_level)
     else: