+++ CHANGES.rst
@@ -3,8 +3,100 @@
 Changes
 #######
 
-0.1.47 (unreleased)
+0.1.51 (unreleased)
 ===================
+
+
+
+0.1.50 (2023-05-22)
+===================
+
+New features:
+
+- Add support for reading data from models with base (`#273`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/273
+
+- Add support for `unique` constraints in tabular manifests (`#148`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/148
+
+Improvements:
+
+- Much better implementation for updating manifest files from SQL as data
+  source (`#346`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/364
+
+- Show better error messages on foreign key constraint errors (`#363`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/363
+
+- Return a non-zero error code if `spinta push` command fails with an error
+  (`#423`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/423
+
+- Add support for older SQLite versions (`#411`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/411
+
+Bug fixes:
+
+- Correctly handle level 3 references, when referenced model does not have a
+  primary key or property references a non-primary key (`#400`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/400
+
+- WIPE command now works on tables with long names (`#431`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/431
+
+
+0.1.49 (2023-04-19)
+===================
+
+Bug fixes:
+
+- Fix issue with order of axes in geometry properties (`#410`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/410
+
+
+- Fix write operations models containing geometry properties (`#417`__,
+  `#418`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/417
+  __ https://github.com/atviriduomenys/spinta/issues/418
+
+
+0.1.48 (2023-04-14)
+===================
+
+Bug fixes:
+
+- Fix issue with dask/pandas version incompatibility (`dask/dask#10164`__).
+
+  __ https://github.com/dask/dask/issues/10164
+
+
+0.1.47 (2023-03-27)
+===================
+
+Improvements:
+
+- Add support for `point(x,y)` and `cast()` functions for sql backend
+  (`#407`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/407
+
+Bug fixes:
+
+- Error when loading manifest from XLSX file, where level is read as integer
+  (`#405`__).
+
+  __ https://github.com/atviriduomenys/spinta/issues/405
+
 
 
 0.1.46 (2023-03-21)